Add initial wifi functionality
WiFi works. If enabled, on boot the 8188fu wifi kernel module will be loaded, wifi chip powered on and network connected automatically. FTP and Telnet access will be available if configured as such. There is currently a very rudimentary "Wifi Setup.pak" in Extras, which can only turn the WiFi on and off on next boot.
Also commented out the bootloader update part from installer as it is not needed for MMP.